e2e: Merge sanity tests
`lightning_routes.ts` contains sanity tests for the lightning routes.
These can be merged with the other sanity tests in `sanity.ts`.
